requirements.txt
# ─────────────────────────────────────────────────────────────
#  XaiSkinLesionResearch — requirements.txt
#
#  Install:
#    pip install -r requirements.txt
#
#  PyTorch (CUDA 12.1 — RTX 3080, WSL2):
#    pip install torch==2.0.1 torchvision==0.15.2 \
#        --index-url https://download.pytorch.org/whl/cu121
# ─────────────────────────────────────────────────────────────

# ── Core ML / Deep Learning ──────────────────────────────────
torch==2.0.1
torchvision==0.15.2

# ── Data & Numerics ──────────────────────────────────────────
numpy<2
pandas==2.2.3
scipy==1.15.2
scikit-learn>=1.3.0

# ── Image Processing ─────────────────────────────────────────
opencv-python==4.10.0.84
Pillow==11.1.0
scikit-image>=0.25.2

# ── Visualisation ────────────────────────────────────────────
matplotlib==3.10.1
seaborn>=0.13.0

# ── XAI Methods ──────────────────────────────────────────────
lime>=0.2.0.1

# ── Progress & Utilities ─────────────────────────────────────
tqdm==4.67.1
tabulate==0.9.0
termcolor==2.5.0
packaging==24.2
cloudpickle==3.1.1

# ── Config ───────────────────────────────────────────────────
PyYAML==6.0.2
omegaconf==2.3.0

# ── Dataset Export ───────────────────────────────────────────
h5py>=3.9.0

# ── Denoising ────────────────────────────────────────────────
bm4d==4.2.5

# ── PDF / Reporting ──────────────────────────────────────────
fpdf==1.7.2
PyMuPDF>=1.23.0

# ── GUI (optional, local viewer) ─────────────────────────────
PySide6==6.7.3

# ── COCO tools ───────────────────────────────────────────────
pycocotools==2.0.8

# ── fvcore / iopath ──────────────────────────────────────────
fvcore==0.1.5.post20221221
iopath==0.1.10

# ── Path utilities ───────────────────────────────────────────
pathlib==1.0.1
